requirements.txt
dash==2.9.1
pandas==1.4.0
openai==0.27.0
plotly==5.9.0
kafka-python==2.0.2

